+++ changelog.md
@@ -1,3 +1,9 @@
+# 2.0.3
+- Reworked Discord Popup.
+- Grind Indicator button moved to the left side of level name in order to improve the compatibility with other mods.
+- Small badge fix.
+- Argon on bootup fix.
+
 # 2.0.2
 - Badges appearing multiple times bugfix.
 - Badges in comments.